Add files via upload
this code is made for cory to use for his camera. code was also updated with function for him and the obstacle detection function that still needs to be tested. Chris has a thomas function that will need to be added to master code that takes tic tracer and knocker readings to give to matrix.